# src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Accounting Node.TS API Specification
  version: 1.1.0
  description: |
    Comprehensive OpenAPI 3.0 specification for the Accounting Node.TS micro-monolith backend.
    Handles multi-tenant financial accounting, customer directories, inventory management,
    transactional invoice generation, and compressed statement streaming.
  contact:
    name: API Support
    email: support@accounting.local
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC

servers:
  - url: http://localhost:420
    description: Local Development Server (Port 420)
  - url: http://localhost:3000
    description: Alternative Local Server (Port 3000)

tags:
  - name: Authentication
    description: User registration, Argon2id credential verification, and JWT issuance
  - name: Customers
    description: Tenant customer management and directory operations
  - name: Inventory
    description: Product catalog and stock management
  - name: Invoices
    description: Transactional invoice generation and historical retrieval
  - name: Statements
    description: Compressed financial statements and audit trail retrieval

paths:
  /auth/signup:
    post:
      tags:
        - Authentication
      summary: Register a new business tenant user
      description: |
        Creates a new user record in the global `users` table.
        Passwords are encrypted using Argon2id. Automatically provisions a JWT token on success.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
            example:
              name: Jane Doe
              username: janedoe
              email: jane@example.com
              password: SecurePassword123!
              gstin: 29ABCDE1234F1Z5
              pan: ABCDE1234F
              aadhaar: 123456789012
              phone: +919876543210
              address: 100 Financial Way, Tech Park, Bangalore
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Username already exists
        '500':
          description: Database transaction or internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Tenant authentication & JWT issuance
      description: Validates username and Argon2id password hash, generating a 1-hour JWT token.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: janedoe
              password: SecurePassword123!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginSuccessResponse'
        '401':
          description: Invalid username or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Invalid password
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /customer/add:
    post:
      tags:
        - Customers
      summary: Create a new customer for a tenant
      description: |
        Inserts a new customer record into `<username>_customers`.
        If the table does not exist, it is automatically created.
        Generates an auto-incremented customer identifier (`cust_001`, `cust_002`, ...).
      operationId: addCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCustomerRequest'
            example:
              name: AcmeIndustries
              address: Plot 42 Industrial Area Phase 2
              phone: 9876543210
              email: accounts@acme.com
              gstIN: 29AAAAA0000A1Z5
              dealer_type: Regular
              pan: AAAAA0000A
              aadhaar: 987654321098
              username: janedoe
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddCustomerResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /customer/get:
    get:
      tags:
        - Customers
      summary: Fetch all customers for a tenant
      description: Retrieves all customer records from `<username>_customers`.
      operationId: getCustomers
      parameters:
        - name: username
          in: query
          required: true
          description: Alphanumeric tenant username
          schema:
            type: string
          example: janedoe
      responses:
        '200':
          description: Customers retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCustomersResponse'
        '422':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /inventory/add:
    post:
      tags:
        - Inventory
      summary: Add an item to tenant inventory
      description: |
        Inserts a product record into `<username>_inventory`.
        Creates the table if it does not yet exist.
      operationId: addInventoryItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProductRequest'
            example:
              username: janedoe
              product_name: Steel Bearings 20mm
              quantity: 200
              unit_price: 45.50
      responses:
        '200':
          description: Product added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  messge:
                    type: string
                    example: Product added
                  result:
                    type: object
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /invoice/add:
    post:
      tags:
        - Invoices
      summary: Create an invoice with lines
      description: |
        Executes a transactional invoice creation:
        1. Generates timestamped transaction ID (`txn_...`).
        2. Inserts header into `<username>_invoices`.
        3. Queries unit price from `<username>_inventory` for each line item.
        4. Calculates total amount per line (`unit_price * quantity`).
        5. Inserts line items into `<username>_invoice_lines`.
        6. Commits transaction or rolls back on any error.
      operationId: createInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequest'
            example:
              username: janedoe
              invoice:
                customer_id: cust_001
                date_time: 2026-09-29T18:00:00.000Z
                total: 4550.00
                total_discount: 100.00
                packaging: 50.00
                freight: 100.00
                taxable_amount: 4550.00
                tax_collected_at_source: 0.00
                round_off: 0.00
                grand_total: 4600.00
                method_of_payment: NEFT
                invoiceLines:
                  - product_id: 1
                    quantity: 100
      responses:
        '201':
          description: Invoice created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateInvoiceResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /invoice/get:
    get:
      tags:
        - Invoices
      summary: Fetch tenant invoices with filtering (Gzip compressed)
      description: |
        Retrieves invoices joined with customer details and line items.
        Applies SQL temporal filter based on `action`.
        Response body is gzip compressed (`Content-Encoding: gzip`).
      operationId: getInvoices
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
          example: janedoe
        - name: action
          in: query
          required: true
          description: Temporal filtering preset
          schema:
            type: string
            enum:
              - today
              - toAndFromDate
              - thisMonth
              - thisWeek
              - thisQuarter
              - thisYear
              - beforeDate
              - afterDate
              - all
          example: thisMonth
        - name: sdate
          in: query
          required: false
          description: Start date ISO string (used when action is toAndFromDate)
          schema:
            type: string
            format: date-time
          example: 2026-09-01T00:00:00.000Z
        - name: endate
          in: query
          required: false
          description: End date ISO string (used when action is toAndFromDate)
          schema:
            type: string
            format: date-time
          example: 2026-09-30T23:59:59.000Z
      responses:
        '200':
          description: Gzip-compressed JSON array of invoice records
          headers:
            Content-Encoding:
              schema:
                type: string
                example: gzip
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceListResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /stmt/get:
    get:
      tags:
        - Statements
      summary: Fetch financial account statement (Gzip compressed)
      description: |
        Retrieves invoices joined with customer profile for audit statements.
        Returns a gzip compressed JSON payload.
      operationId: getStatements
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
          example: janedoe
        - name: action
          in: query
          required: true
          schema:
            type: string
            enum:
              - today
              - toAndFromDate
              - thisMonth
              - thisWeek
              - thisQuarter
              - thisYear
              - beforeDate
              - afterDate
              - all
          example: thisYear
        - name: sdate
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: endate
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Statement records delivered with gzip compression
          headers:
            Content-Encoding:
              schema:
                type: string
                example: gzip
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceListResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    SignupRequest:
      type: object
      required:
        - name
        - username
        - email
        - password
        - gstin
        - pan
        - aadhaar
        - phone
        - address
      properties:
        name:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        gstin:
          type: string
        pan:
          type: string
        aadhaar:
          type: string
        phone:
          type: string
        address:
          type: string

    AuthSuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: User created successfully
        user:
          type: object
          properties:
            name:
              type: string
            username:
              type: string
            email:
              type: string
            gstIn:
              type: string
            pan_card:
              type: string
            adhaar:
              type: string
            phone:
              type: string
            address:
              type: string
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string

    LoginSuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: Login successful
        token:
          type: string

    AddCustomerRequest:
      type: object
      required:
        - name
        - address
        - phone
        - email
        - gstIN
        - pan
        - aadhaar
        - username
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        gstIN:
          type: string
        dealer_type:
          type: string
        pan:
          type: string
        aadhaar:
          type: string
        username:
          type: string

    CustomerRecord:
      type: object
      properties:
        id:
          type: integer
        cust_id:
          type: string
        name:
          type: string
        address:
          type: string
        phone_number:
          type: string
        email:
          type: string
        gstin:
          type: string
        dealer_type:
          type: string
        pan_card:
          type: string
        aadhaar:
          type: string

    AddCustomerResponse:
      type: object
      properties:
        message:
          type: string
          example: customer created successfully
        customer:
          type: array
          items:
            $ref: '#/components/schemas/CustomerRecord'

    GetCustomersResponse:
      type: object
      properties:
        message:
          type: string
          example: customer fetched successfully
        customers:
          type: array
          items:
            $ref: '#/components/schemas/CustomerRecord'

    AddProductRequest:
      type: object
      required:
        - username
        - product_name
        - quantity
        - unit_price
      properties:
        username:
          type: string
        product_name:
          type: string
        quantity:
          type: integer
        unit_price:
          type: number
          format: float

    InvoiceLineInput:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: integer
        quantity:
          type: integer

    InvoiceInputPayload:
      type: object
      required:
        - customer_id
        - method_of_payment
        - invoiceLines
      properties:
        customer_id:
          type: string
        date_time:
          type: string
          format: date-time
        total:
          type: number
        total_discount:
          type: number
        packaging:
          type: number
        freight:
          type: number
        taxable_amount:
          type: number
        tax_collected_at_source:
          type: number
        round_off:
          type: number
        grand_total:
          type: number
        method_of_payment:
          type: string
        invoiceLines:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceLineInput'

    CreateInvoiceRequest:
      type: object
      required:
        - username
        - invoice
      properties:
        username:
          type: string
        invoice:
          $ref: '#/components/schemas/InvoiceInputPayload'

    CreateInvoiceResponse:
      type: object
      properties:
        message:
          type: string
          example: invoice created successfully
        invoice:
          type: object

    InvoiceLineDetail:
      type: object
      properties:
        contents:
          type: string
        unit_price:
          type: number
        quantity:
          type: integer
        amount:
          type: number

    InvoiceDetail:
      type: object
      properties:
        id:
          type: integer
        transaction_id:
          type: string
        customer_id:
          type: object
        date_time:
          type: string
          format: date-time
        total:
          type: number
        total_discount:
          type: number
        packaging:
          type: number
        freight:
          type: number
        taxable_amount:
          type: number
        tax_collected_at_source:
          type: number
        round_off:
          type: number
        grand_total:
          type: number
        method_of_payment:
          type: string
        lines:
          type: array
          items:
            type: array
            items:
              $ref: '#/components/schemas/InvoiceLineDetail'

    InvoiceListResponse:
      type: object
      properties:
        invoices:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceDetail'

    ValidationErrorResponse:
      type: object
      properties:
        Errors:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              msg:
                type: string
              path:
                type: string
              location:
                type: string

    ErrorResponse:
      type: object
      properties:
        error:
          type: string